requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cloudinary==1.44.1
Django==5.2.6
django-htmx==1.26.0
pillow==11.3.0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
